// acro_frontend/src/components/SiderTabs/RelatedVideos/interface.ts
export interface UserCard {
  user_id?: number;
  username?: string;
  nickname?: string;
  reg_time?: string;

  follow_count?: number;
  be_followed?: boolean;
  be_followed_count?: number;

  be_liked_count?: number;
  be_favorite_count?: number;
  be_commented_count?: number;
  be_forwarded_count?: number;
  be_watched_count?: number;

  avatar_url?: string;
}

export interface VideoCard {
  user?: UserCard;
  video_uid?: string;
  type?: number;
  content?: string;
  keyword?: string;
  upload_time?: string;

  be_liked_count?: number;
  be_favorite_count?: number;
  be_commented_count?: number;
  be_forwarded_count?: number;
  be_watched_count?: number;

  is_user_liked: boolean;
  is_user_favorite: boolean;
  is_user_uploaded: boolean;
  is_user_watched: boolean;
  is_user_last_play: boolean;
  cover_url?: string;
  play_url?: string;

  is_deleted?: boolean;
  is_top?: boolean;
  is_private?: boolean;
}